+++ src/app/entities/IncomingInvoiceProduct.entity.ts
@@ -23,11 +23,11 @@ const IncomingInvoiceProduct = sequelize.define(
         allowNull: false,
       },
       price: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.FLOAT(11, 2),
         allowNull: false,
       },
       quantity: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
       date: {
